requirements.txt
aniso8601==9.0.1
ansible==7.0.0
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
autoflake8
Automat==0.7.0
bcrypt==3.1.6
CacheControl==0.12.5
cachetools==5.2.0
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloud-init-seed==0.3.0
colorama==0.4.6
configobj==5.0.6
constantly==15.1.0
cryptography==41.0.2
defusedxml==0.5.0
dnspython==1.15.0
docutils==0.14

ecdsa==0.13.3
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Injector==0.15.0
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
flask_sqlalchemy==3.0.5
google-auth==2.10.0
graphviz==0.9
humanfriendly==10.0
hyperlink==18.0.0

importlib-metadata==5.2.0
incremental==17.5.0
ipaddr==2.2.0
itsdangerous==2.0.1

Jinja2==3.0.1
jmespath==0.9.3
Js2Py==0.66
jsonpatch==1.23
jsonpointer==2.0
jsonschema==2.6.0
keyring==23.5.0
kubernetes==24.2.0

lockfile==0.12.2
lxml==4.9.2
m2r==0.2.0
MarkupSafe==2.0.1
mistune==0.8.3

msgpack==0.5.6
netaddr==0.7.19

ntplib==0.4.0
oauthlib==3.2.2
packaging==22.0
paramiko==2.7.2
pexpect==4.6.0
pillow==9.5.0
ply==3.11
polling==0.5.0
prettytable==0.7.2
prompt-toolkit==2.0.9
PTable==0.9.2

ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21

pycryptodome==3.16.0

pydantic==1.9.0

Pygments==2.15.0
PyHamcrest==1.9.0
pyjsparser==2.7.1

pyjwt==2.8.0
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
pytenable
python-dateutil==2.8.2
python-pam==1.8.2
python-terraform==0.10.1
pytz==2018.5
pyvcloud==23.0.4
pyvim==2.0.22
pyvmomi==6.7.0.2018.9
pywbem==0.15.0
PyYAML==6.0

requests==2.28.1
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.9
ruamel.yaml==0.17.16


ruamel.yaml.clib==0.2.6
SecretStorage==3.3.3
service-identity==18.1.0
setuptools-scm==3.1.0

simplejson==3.16.0
six==1.16.0
tabulate==0.9.0
tqdm==4.62.3



Twisted==22.10.0rc1
typing==3.6.6
typing_extensions==4.4.0
tzlocal==1.5.1
unittest-xml-reporting==3.2.0
urllib3==1.26.17
vcd-api-schemas-type==10.3.0.dev72
vcd-cli==24.1.0

waitress==2.1.2


wcwidth==0.1.7
WebOb==1.8.2
websocket-client==1.3.3
Werkzeug==2.2.3
zipp==3.11.0
zope.interface==5.1.0
